Updated examples, added support for getting an existing S3 bucket
