+++ packages/s3-deploy/lib/rootSite.ts
@@ -1,21 +1,10 @@
 #!/usr/bin/env node
-import * as s3 from 'aws-cdk-lib/aws-s3';
 import * as cloudfront from 'aws-cdk-lib/aws-cloudfront';
 import * as cloudfront_origins from 'aws-cdk-lib/aws-cloudfront-origins';
 import * as iam from 'aws-cdk-lib/aws-iam';
-import { CfnOutput, RemovalPolicy } from 'aws-cdk-lib';
+import { CfnOutput } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
-
-const cors = [
-  {
-    allowedMethods: [
-      s3.HttpMethods.GET,
-      s3.HttpMethods.HEAD,
-    ],
-    allowedOrigins: ['*'],
-    allowedHeaders: ['*'],
-  },
-]
+import getBucket from './getBucket.js';
 
 /**
  * Constructs an S3 bucket definition for dicomweb data, for use by distribution.
@@ -25,15 +14,7 @@ const rootSite = function(site: Construct, name: string, cloudfrontOAI: cloudfro
 
   const { Bucket: dicomwebName } = props;
 
-  const dicomwebBucket = new s3.Bucket(site, dicomwebName, {
-    bucketName: dicomwebName,
-    cors,
-    websiteIndexDocument: 'index.json',
-    websiteErrorDocument: 'error.html',
-    publicReadAccess: true,
-    removalPolicy: RemovalPolicy.DESTROY, // NOT recommended for production code
-    autoDeleteObjects: true, // NOT recommended for production code
-  });
+  const dicomwebBucket = getBucket(site, dicomwebName, props);
   new CfnOutput(site, 'DICOMweb BucketURL', { value: dicomwebBucket.bucketWebsiteUrl});
 
   dicomwebBucket.addToResourcePolicy(new iam.PolicyStatement({
